Move internally used types to dedicated classes
