+++ src/auth/auth-api-request.ts
@@ -19,10 +19,11 @@ import * as validator from '../utils/validator';
 
 import {deepCopy, deepExtend} from '../utils/deep-copy';
 import {
-  UserIdentifier, isUidIdentifier, isEmailIdentifier, isPhoneIdentifier,
-  isProviderIdentifier, UidIdentifier, EmailIdentifier, PhoneIdentifier,
+  UserIdentifier, UidIdentifier, EmailIdentifier, PhoneIdentifier,
   ProviderIdentifier,
 } from './identifier';
+import { isUidIdentifier, isEmailIdentifier, isPhoneIdentifier,
+  isProviderIdentifier} from './identifier-internal';
 import {FirebaseApp} from '../firebase-app';
 import {AuthClientErrorCode, FirebaseAuthError} from '../utils/error';
 import {
@@ -30,16 +31,19 @@ import {
 } from '../utils/api-request';
 import {CreateRequest, UpdateRequest} from './user-record';
 import {
-  UserImportBuilder, UserImportOptions, UserImportRecord,
-  UserImportResult, AuthFactorInfo, convertMultiFactorInfoToServerFormat,
+  UserImportOptions, UserImportRecord, UserImportResult
 } from './user-import-builder';
+import {UserImportBuilder, AuthFactorInfo, convertMultiFactorInfoToServerFormat} from './user-import-builder-internal';
 import * as utils from '../utils/index';
 import {ActionCodeSettings, ActionCodeSettingsBuilder} from './action-code-settings-builder';
 import {
-  SAMLConfig, OIDCConfig, OIDCConfigServerResponse, SAMLConfigServerResponse,
-  OIDCConfigServerRequest, SAMLConfigServerRequest, AuthProviderConfig,
+  AuthProviderConfig,
   OIDCUpdateAuthProviderRequest, SAMLUpdateAuthProviderRequest,
 } from './auth-config';
+import {
+  SAMLConfig, OIDCConfig, OIDCConfigServerResponse, SAMLConfigServerResponse,
+  OIDCConfigServerRequest, SAMLConfigServerRequest
+} from './auth-config-internal';
 import {Tenant, TenantOptions, TenantServerResponse} from './tenant';
