+++ test/unit/auth/auth-config.spec.ts
@@ -20,13 +20,17 @@ import * as sinonChai from 'sinon-chai';
 import * as chaiAsPromised from 'chai-as-promised';
 
 import {deepCopy} from '../../../src/utils/deep-copy';
+
+import {
+  SAMLUpdateAuthProviderRequest,
+  OIDCUpdateAuthProviderRequest, SAMLAuthProviderConfig, OIDCAuthProviderConfig,
+} from '../../../src/auth/auth-config';
+
 import {
   OIDCConfig, SAMLConfig, SAMLConfigServerRequest,
   SAMLConfigServerResponse, OIDCConfigServerRequest,
-  OIDCConfigServerResponse, SAMLUpdateAuthProviderRequest,
-  OIDCUpdateAuthProviderRequest, SAMLAuthProviderConfig, OIDCAuthProviderConfig,
-  EmailSignInConfig,
-} from '../../../src/auth/auth-config';
+  OIDCConfigServerResponse, EmailSignInConfig,
+} from '../../../src/auth/auth-config-internal';
 
 
 chai.should();
