+++ test/unit/auth/token-verifier.spec.ts
@@ -28,10 +28,10 @@ import * as chaiAsPromised from 'chai-as-promised';
 import LegacyFirebaseTokenGenerator = require('firebase-token-generator');
 
 import * as mocks from '../../resources/mocks';
-import {FirebaseTokenGenerator, ServiceAccountSigner} from '../../../src/auth/token-generator';
-import * as verifier from '../../../src/auth/token-verifier';
+import {FirebaseTokenGenerator, ServiceAccountSigner} from '../../../src/auth/token-generator-internal';
+import * as verifier from '../../../src/auth/token-verifier-internal';
 
-import {ServiceAccountCredential} from '../../../src/auth/credential';
+import {ServiceAccountCredential} from '../../../src/auth/credential-internal';
 import { AuthClientErrorCode } from '../../../src/utils/error';
 import { FirebaseApp } from '../../../src/firebase-app';
